+++ internal/slack/export_test.go
@@ -1,13 +1,5 @@
 package slack
 
-func SetSlackFilesUploadURL(u string) (resetFunc func()) {
-	var tmp string
-	tmp, slackFilesUploadURL = slackFilesUploadURL, u
-	return func() {
-		slackFilesUploadURL = tmp
-	}
-}
-
 func SetFilesGetUploadURLExternalURL(u string) (resetFunc func()) {
 	var tmp string
 	tmp, filesGetUploadURLExternalURL = filesGetUploadURLExternalURL, u